requirements.txt
openai>=1.30.0
PyGithub>=2.4.0
requests>=2.32.0
requests-oauthlib>=1.3.1
tweepy>=4.14.0
pymupdf4llm>=0.0.17
weasyprint>=62.0
jinja2>=3.1.4
markdown>=3.6
python-dotenv>=1.0.0
tenacity>=8.3.0
loguru>=0.7.2
pydantic>=2.7.0
